Provide more detailed error information for load.
When attempting to load a PNG file, if that file is unsupported
it is useful to provide the developer with a more precise
error message indicating the exact cause of the error.

Have also changed the type of exception to be a ValueError rather
than a TypeError. In this case the type is correct, however the
value recevied is incorrect (or at least unsupported).

Probably worth changing to a specific exception in the future.