Add a reference RGB only image and associated test
Support for RGB only (as opposed to RGBA) images is desired.
This provides a baseline to test support against.

The assumption is that when loading an RGB image it will be
converted to the internal RGBA format. This provides a good
baseline of support without needing to change other parts
of the module.

In the future the color type (RGB vs RGBA) may be supported
directly and preserved with a round-trip, however that is
beyond the scope of initial support.